🚚 change file structure for account page & product all list page
