+++ src/app/(main)/accounts/fund/FundPageTab.tsx
@@ -1,0 +1,10 @@
+import { PurchaseHistoryItemWrapContainer } from '@components/product/ProductItem';
+
+// 펀드 페이지 탭
+export default function FundPageTab() {
+  return (
+    <>
+      <PurchaseHistoryItemWrapContainer />
+    </>
+  );
+}
