+++ src/app/(main)/accounts/page.tsx
@@ -1,10 +1,19 @@
-import { PurchaseHistoryTab } from '@components/tab/categoryTab';
+import { PurchaseHistoryItemWrapContainer } from '@components/product/ProductItem';
 
 export default function Account() {
   return (
     <>
       {/* 구매내역이 디폴트 */}
       <PurchaseHistoryTab />
+    </>
+  );
+}
+
+// 구매내역 탭
+function PurchaseHistoryTab() {
+  return (
+    <>
+      <PurchaseHistoryItemWrapContainer />
     </>
   );
 }
